# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The GoMicro service is a template for microservices written in go.",
        "title": "GoMicro service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/config": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configs"
                ],
                "summary": "Get a store for a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response with config as json",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apiv1.ConfigDescription"
                            }
                        }
                    },
                    "400": {
                        "description": "client error information as json",
                        "schema": {
                            "$ref": "#/definitions/serror.Serr"
                        }
                    },
                    "500": {
                        "description": "server error information as json",
                        "schema": {
                            "$ref": "#/definitions/serror.Serr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configs"
                ],
                "summary": "Create a new store for a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add store",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "tenant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "client error information as json",
                        "schema": {
                            "$ref": "#/definitions/serror.Serr"
                        }
                    },
                    "500": {
                        "description": "server error information as json",
                        "schema": {
                            "$ref": "#/definitions/serror.Serr"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configs"
                ],
                "summary": "Delete a store for a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "client error information as json",
                        "schema": {
                            "$ref": "#/definitions/serror.Serr"
                        }
                    }
                }
            }
        },
        "/config/size": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configs"
                ],
                "summary": "Get size of a store for a tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "size",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "client error information as json",
                        "schema": {
                            "$ref": "#/definitions/serror.Serr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apiv1.ConfigDescription": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                },
                "storeid": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
        "serror.Serr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "apikey",
            "in": "header"
        }
    }
}